[tools/tests/test_main.c] Verify fork() support
The tests will not pass if libcheck is compiled with CF_FORK=no
because they have been written assuming CF_FORK=yes. Each test
inherits a pristine copy of global state in the forked child process.
Mutations of global state by the child process are not propagated
forward to the next test.

This commit exposes the implicit assumption of CF_FORK=yes as
an explicit test.